更新至 v1.4.0 並改進多項功能
將專案版本從 1.2.0 更新至 1.4.0。
在 `MySqlSchemaReader.cs` 中新增 `using System.Reflection.PortableExecutable;`。
多處新增 `reader.Close();` 以確保資料讀取器在使用後被正確關閉。
將 `column.MaxLength` 型別從 `int` 改為 `long`。
修改 `ReadForeignKeyRules` 方法，移除 `MySqlConnection` 參數，並在方法內部建立連線。
在 `Options.cs` 中，將 `config` 選項的預設值設為 `appsettings.json`，並移除 `settings` 和 `init` 選項。
在 `Program.cs` 中，移除初始化配置文件的邏輯，並更新配置文件的讀取邏輯。
新增 `launchSettings.json` 文件，包含 WSL2 的配置。
在 `TableDefinition.cs` 中，將 `MaxLength` 屬性的型別從 `int?` 改為 `long?`。
更新 `README.md` 文件，詳細說明工具的功能、安裝步驟、使用方法和配置選項，並提供生成代碼的範例。
移除 `README.md` 中的 `init` 命令說明，並更新配置文件的範例。
更新 `README.md` 中的發佈歷史，新增 v1.0.0 版本的說明。

+++ CHC.EF.Reverse.ConsoleApp/TableDefinition.cs
@@ -135,7 +135,7 @@ namespace CHC.EF.Reverse.ConsoleApp
         public bool IsIdentity { get; set; }
         public bool IsComputed { get; set; }
         public string DefaultValue { get; set; }
-        public int? MaxLength { get; set; }
+        public long? MaxLength { get; set; }
         public int? Precision { get; set; }
         public int? Scale { get; set; }
         public string Comment { get; set; }
